Added SMUX filtering to snmpd.conf.erb
On Linux SMUX is disabled or not compiled in by default. On FreeBSD this is not
the case, and snmpd listens for SMUX connections on port 199/tcp on all
interfaces. As this is a security risk made SMUX listen on 127.0.0.1 only. There
doesn't seem to be any IPv6 support in the SMUX part of snmpd yet.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>