Do not use iniface in packet filtering rules unless defined
Previous the primary network interface was added to the firewall rule. This may
work ok in environments such as EC2 VPCs where you tend to have just one
interface which is in the private network. But it breaks horribly in Hetzner,
for example, where the private interface is actually not the primary interface.

Signed-off-by: Samuli Seppänen <samuli.seppanen@puppeteers.fi>